+++ mavis/test/__init__.py
@@ -31,6 +31,7 @@ from .fixtures import (
     nurse,
     onboard_and_delete,
     onboarding,
+    online_consent_page,
     organisation,
     programmes_enabled,
     programmes_page,
@@ -83,6 +84,7 @@ __all__ = [
     "nurse",
     "onboard_and_delete",
     "onboarding",
+    "online_consent_page",
     "organisation",
     "programmes_page",
     "programmes_enabled",
