+++ mavis/test/fixtures/__init__.py
@@ -41,6 +41,7 @@ from .pages import (
     import_records_page,
     log_in_page,
     match_consent_response_page,
+    online_consent_page,
     programmes_page,
     review_school_move_page,
     school_moves_page,
@@ -88,6 +89,7 @@ __all__ = [
     "nurse",
     "onboard_and_delete",
     "onboarding",
+    "online_consent_page",
     "organisation",
     "programmes_page",
     "programmes_enabled",
