+++ mavis/test/pages/verbal_consent.py
@@ -1,14 +1,10 @@
-from datetime import date
 from playwright.sync_api import Page, expect
-from typing import Optional
 
 from mavis.test.models import (
     ConsentMethod,
     ConsentRefusalReason,
     Programme,
-    School,
     Parent,
-    Child,
     ConsentOption,
 )
 from mavis.test.annotations import step
@@ -18,53 +14,15 @@ class VerbalConsentPage:
     def __init__(self, page: Page):
         self.page = page
 
-        self.first_name_textbox = self.page.get_by_role("textbox", name="First name")
-        self.last_name_textbox = self.page.get_by_role("textbox", name="Last name")
         self.yes_radio = self.page.get_by_role("radio", name="Yes", exact=True)
         self.no_radio = self.page.get_by_role("radio", name="No", exact=True)
-        self.preferred_first_name_textbox = self.page.get_by_role(
-            "textbox", name="Preferred first name (optional)"
-        )
-        self.preferred_last_name_textbox = self.page.get_by_role(
-            "textbox", name="Preferred last name (optional)"
-        )
         self.continue_button = self.page.get_by_role("button", name="Continue")
-        self.dob_day_textbox = self.page.get_by_role("textbox", name="Day")
-        self.dob_month_textbox = self.page.get_by_role("textbox", name="Month")
-        self.dob_year_textbox = self.page.get_by_role("textbox", name="Year")
-        self.displayed_school_name = self.page.get_by_test_id("school-name")
-        self.confirm_school_radio = self.page.get_by_role(
-            "radio", name="Yes, they go to this school"
-        )
-        self.select_different_school_radio = self.page.get_by_role(
-            "radio", name="No, they go to a different school"
-        )
-        self.school_name_combobox = self.page.get_by_role("combobox")
-        self.full_name_textbox = self.page.get_by_role("textbox", name="Full name")
-        self.email_address_textbox = self.page.get_by_role(
-            "textbox", name="Email address"
-        )
         self.phone_textbox = self.page.get_by_role("textbox", name="Phone number")
         self.text_alerts_checkbox = self.page.get_by_role(
             "checkbox", name="Get updates by text message"
         )
         self.mobile_only_text_checkbox = self.page.get_by_role(
             "checkbox", name="I can only receive text"
-        )
-        self.mobile_only_voice_checkbox = self.page.get_by_role(
-            "checkbox", name="I can only receive voice calls"
-        )
-        self.address_line_1_textbox = self.page.get_by_role(
-            "textbox", name="Address line 1"
-        )
-        self.address_line_2_textbox = self.page.get_by_role(
-            "textbox", name="Address line 2 (optional)"
-        )
-        self.address_city_textbox = self.page.get_by_role(
-            "textbox", name="Town or city"
-        )
-        self.address_postcode_textbox = self.page.get_by_role(
-            "textbox", name="Postcode"
         )
         self.give_details_textbox = self.page.get_by_role(
             "textbox", name="Give details"
@@ -77,9 +35,6 @@ class VerbalConsentPage:
             method: self.page.get_by_role("radio", name=method)
             for method in ConsentMethod
         }
-        self.change_phone_link = self.page.get_by_role(
-            "link", name="Change   your phone"
-        )
         self.add_phone_number_link = self.page.get_by_role(
             "link", name="Add phone number"
         )
@@ -104,30 +59,6 @@ class VerbalConsentPage:
         self.yes_they_agree_radio = self.page.get_by_role(
             "radio", name="Yes, they agree"
         )
-
-        self.doubles_consent_both_radio = self.page.get_by_role(
-            "radio", name="Yes, I agree to them having"
-        )
-        self.doubles_consent_one_radio = self.page.get_by_role(
-            "radio", name="I agree to them having one of"
-        )
-        self.doubles_consent_menacwy_radio = self.page.get_by_role(
-            "radio", name="MenACWY"
-        )
-        self.doubles_consent_tdipv_radio = self.page.get_by_role("radio", name="Td/IPV")
-        self.flu_agree_injection_radio = self.page.get_by_role(
-            "radio", name="Yes, I agree to the alternative flu injection"
-        )
-        self.flu_agree_nasal_radio = self.page.get_by_role(
-            "radio", name="Yes, I agree to them having the nasal spray vaccine"
-        )
-        self.hpv_consent_agree_radio = self.page.get_by_role(
-            "radio", name="Yes, I agree"
-        )
-        self.no_consent_radio = self.page.get_by_role("radio", name="No")
-        self.sessions_link = self.page.get_by_label("Primary navigation").get_by_role(
-            "link", name="Sessions"
-        )
         self.online_flu_agree_nasal_radio = self.page.get_by_role(
             "radio", name="Yes, for the nasal spray"
         )
@@ -148,121 +79,9 @@ class VerbalConsentPage:
         self.phone_textbox.fill(phone)
         self.text_alerts_checkbox.check()
 
-    @step("Fill child name details")
-    def fill_child_name_details(
-        self,
-        first_name: str,
-        last_name: str,
-        known_as_first: Optional[str] = None,
-        known_as_last: Optional[str] = None,
-    ) -> None:
-        self.first_name_textbox.fill(first_name)
-        self.last_name_textbox.fill(last_name)
-
-        if known_as_first or known_as_last:
-            self.yes_radio.click()
-            self.preferred_first_name_textbox.fill(str(known_as_first))
-            self.preferred_last_name_textbox.fill(str(known_as_last))
-        else:
-            self.no_radio.click()
-
-        self.click_continue()
-
-    @step("Fill child date of birth")
-    def fill_child_date_of_birth(self, date_of_birth: date) -> None:
-        self.dob_day_textbox.fill(str(date_of_birth.day))
-        self.dob_month_textbox.fill(str(date_of_birth.month))
-        self.dob_year_textbox.fill(str(date_of_birth.year))
-        self.click_continue()
-
-    @step("Select child school")
-    def select_child_school(self, school: School) -> None:
-        name = str(school)
-
-        if name == self.displayed_school_name.inner_text().strip():
-            self.confirm_school_radio.click()
-        else:
-            self.select_different_school_radio.click()
-            self.click_continue()
-
-            self.page.wait_for_load_state()
-
-            self.school_name_combobox.fill(name)
-            self.page.get_by_role("option", name=name).click()
-
-        self.click_continue()
-
     @step("Click on {1} radio button")
     def click_radio_button(self, name: str) -> None:
         self.page.get_by_role("radio", name=name).check()
-
-    @step("Fill email address")
-    def fill_email_address(self, email):
-        self.email_address_textbox.fill(email)
-
-    @step("Fill parent name")
-    def fill_parent_name(self, parent_name: str) -> None:
-        self.full_name_textbox.fill(parent_name)
-
-    def fill_parent_details(
-        self,
-        parent: Parent,
-        phone: Optional[str] = None,
-    ) -> None:
-        self.fill_parent_name(parent.full_name)
-        self.click_radio_button(parent.relationship)
-
-        if parent.email_address:
-            self.fill_email_address(parent.email_address)
-
-        if phone:
-            self.fill_phone_number_and_receive_text_alerts(phone)
-
-        self.click_continue()
-
-    @step("Agree to doubles vaccinations: {1}")
-    def agree_to_doubles_vaccinations(self, *programmes: Programme):
-        if Programme.MENACWY in programmes and Programme.TD_IPV in programmes:
-            self.doubles_consent_both_radio.check()
-        elif Programme.MENACWY in programmes:
-            self.doubles_consent_one_radio.check()
-            self.doubles_consent_menacwy_radio.check()
-        elif Programme.TD_IPV in programmes:
-            self.doubles_consent_one_radio.check()
-            self.doubles_consent_tdipv_radio.check()
-        self.click_continue()
-
-    @step("Agree to Flu vaccination (consent_option = {consent_option})")
-    def agree_to_flu_vaccination(self, consent_option: ConsentOption):
-        if consent_option is ConsentOption.INJECTION:
-            self.flu_agree_injection_radio.check()
-        else:
-            self.flu_agree_nasal_radio.check()
-        self.click_continue()
-        if consent_option is ConsentOption.BOTH:
-            self.answer_yes()
-        elif consent_option is ConsentOption.NASAL_SPRAY:
-            self.answer_no()
-
-    @step("Agree to HPV vaccination")
-    def agree_to_hpv_vaccination(self):
-        self.hpv_consent_agree_radio.check()
-        self.click_continue()
-
-    @step("Don't agree to vaccination")
-    def dont_agree_to_vaccination(self):
-        self.no_consent_radio.check()
-        self.click_continue()
-
-    @step("Fill address details")
-    def fill_address_details(
-        self, line1: str, line2: str, city: str, postcode: str
-    ) -> None:
-        self.address_line_1_textbox.fill(line1)
-        self.address_line_2_textbox.fill(line2)
-        self.address_city_textbox.fill(city)
-        self.address_postcode_textbox.fill(postcode)
-        self.click_continue()
 
     def set_all_health_questions_to_no(
         self,
@@ -330,10 +149,6 @@ class VerbalConsentPage:
     def click_gillick_competent(self):
         self.child_gillick_competent_radio.check()
 
-    @step("Click on Sessions")
-    def click_sessions(self):
-        self.sessions_link.click()
-
     @step("Click on Yes, for the nasal spray")
     def click_yes_for_nasal_spray(self):
         self.online_flu_agree_nasal_radio.check()
@@ -342,32 +157,8 @@ class VerbalConsentPage:
     def click_yes_for_injected_vaccine(self):
         self.online_flu_agree_injection_radio.check()
 
-    def answer_yes(self, details: Optional[str] = None):
-        self.select_yes()
-        if details:
-            self.give_details(details)
-        self.click_continue()
-
-    def answer_no(self):
-        self.select_no()
-        self.click_continue()
-
-    def select_consent_not_given_reason(
-        self, reason: ConsentRefusalReason, details: Optional[str] = None
-    ) -> None:
-        self.click_consent_refusal_reason(reason)
-        if reason.requires_details:
-            self.click_continue()
-            self.give_details(str(details))
-
-        self.click_continue()
-
     def expect_text_in_alert(self, text: str) -> None:
         expect(self.page.get_by_role("alert")).to_contain_text(text)
-
-    def expect_confirmation_text(self, text: str) -> None:
-        confirmation = self.page.locator(".nhsuk-panel.nhsuk-panel--confirmation")
-        expect(confirmation).to_contain_text(text)
 
     def change_parent_phone(self):
         self.fill_phone_number_and_receive_text_alerts("7700900000")
@@ -495,68 +286,3 @@ class VerbalConsentPage:
         self.click_continue()
         self.click_safe_to_vaccinate(programme=programme, consent_option=consent_option)
         self.click_continue()
-
-    def go_to_url(self, url: str):
-        self.page.goto(url)
-
-    def fill_details(
-        self,
-        child: Child,
-        parent: Parent,
-        schools: list[School],
-        change_school: bool = False,
-    ):
-        self.fill_child_name_details(*child.name)
-        self.fill_child_date_of_birth(child.date_of_birth)
-
-        if change_school:
-            self.select_child_school(schools[1])
-        else:
-            self.select_child_school(schools[0])
-
-        self.fill_parent_details(parent)
-
-    def answer_health_questions(self, number_of_questions: int, health_question: bool):
-        for _ in range(number_of_questions):
-            if health_question:
-                self.answer_yes("More details")
-            else:
-                self.answer_no()
-
-    def check_final_consent_message(
-        self,
-        child: Child,
-        programmes: list[Programme],
-        health_question: bool,
-        consent_option: ConsentOption = ConsentOption.INJECTION,
-    ):
-        def programme_display(programme, consent_option) -> str:
-            if programme == Programme.FLU:
-                return (
-                    "injected flu"
-                    if consent_option is ConsentOption.INJECTION
-                    else "nasal spray flu"
-                )
-            else:
-                return str(programme)
-
-        programmes_str = " and ".join(
-            programme_display(programme, consent_option) for programme in programmes
-        )
-
-        if programmes == [Programme.MENACWY] or programmes == [Programme.TD_IPV]:
-            title = f"Consent for the {programmes_str} vaccination confirmed"
-        else:
-            title = "Consent confirmed"
-
-        if health_question:
-            body = (
-                f" As you answered ‘yes’ to some of the health questions, "
-                f"we need to check the {programmes_str} vaccination{'s are' if len(programmes) > 1 else ' is'} suitable for "
-                f"{child.first_name} {child.last_name}. We’ll review your answers and get in touch again soon."
-            )
-        else:
-            body = f"{child.first_name} {child.last_name} is due to get the {programmes_str} vaccination{'s' if len(programmes) > 1 else ''} at school"
-
-        final_message = "".join([title, body])
-        self.expect_confirmation_text(final_message)
