+++ tests/test_online_consent_hpv.py
@@ -16,18 +16,18 @@ def start_consent(url, page, start_page):
     start_page.start()
 
 
-def test_refused(verbal_consent_page, schools, children):
+def test_refused(online_consent_page, schools, children):
     child = children[Programme.HPV][0]
     schools = schools[Programme.HPV]
 
-    verbal_consent_page.fill_details(child, child.parents[0], schools)
-    verbal_consent_page.dont_agree_to_vaccination()
-    verbal_consent_page.select_consent_not_given_reason(
+    online_consent_page.fill_details(child, child.parents[0], schools)
+    online_consent_page.dont_agree_to_vaccination()
+    online_consent_page.select_consent_not_given_reason(
         reason=ConsentRefusalReason.VACCINE_ALREADY_RECEIVED,
         details="Vaccine already received in previous school",
     )
-    verbal_consent_page.click_confirm()
-    verbal_consent_page.expect_confirmation_text(
+    online_consent_page.click_confirm()
+    online_consent_page.expect_confirmation_text(
         f"Consent refusedYou’ve told us that you do not want {child.first_name} {child.last_name} to get the HPV vaccination at school"
     )
 
@@ -39,7 +39,7 @@ def test_refused(verbal_consent_page, schools, children):
     "health_question", (False, True), ids=lambda v: f"health_question: {v}"
 )
 def test_given(
-    verbal_consent_page,
+    online_consent_page,
     schools,
     change_school,
     health_question,
@@ -48,11 +48,11 @@ def test_given(
     child = children[Programme.HPV][0]
     schools = schools[Programme.HPV]
 
-    verbal_consent_page.fill_details(child, child.parents[0], schools, change_school)
-    verbal_consent_page.agree_to_hpv_vaccination()
-    verbal_consent_page.fill_address_details(*child.address)
-    verbal_consent_page.answer_health_questions(4, health_question=health_question)
-    verbal_consent_page.click_confirm()
-    verbal_consent_page.check_final_consent_message(
+    online_consent_page.fill_details(child, child.parents[0], schools, change_school)
+    online_consent_page.agree_to_hpv_vaccination()
+    online_consent_page.fill_address_details(*child.address)
+    online_consent_page.answer_health_questions(4, health_question=health_question)
+    online_consent_page.click_confirm()
+    online_consent_page.check_final_consent_message(
         child, programmes=[Programme.HPV], health_question=health_question
     )
